Planilla de Suelto y tabla de registrados

+++ src/Controlador/Producto.java
@@ -15,7 +15,7 @@ public class Producto {
     public String marca;
     public String modelo;
     public String proveedor;
-    public int precio;
+    public float precio;
     public int cantidad;
     
         public Producto (){
